update_extern_crate belongs to CrateMetadata
This also allows us to make extern_crate field private
